fix: 댓글 삭제 실패 여부 수정

+++ be-spring/iscode/iscode/src/main/java/com/ssafy/iscode/rcomment/model/dao/ReviewCommentRepository.java
@@ -50,6 +50,9 @@ public class ReviewCommentRepository {
 
     public int deleteComment(ReviewComment reviewComment){
         try{
+            if(reviewComment == null){
+                return 0;
+            }
             if(reviewComment.getChildren().isEmpty()){
                em.remove(reviewComment);
             }
